# Conflicts:
#	cookiecutter/__init__.py
#	cookiecutter/generate.py
#	cookiecutter/main.py
#	setup.cfg
#	tests/test_main.py

+++ cookiecutter/generate.py
@@ -1,31 +1,27 @@
-# -*- coding: utf-8 -*-
-
 """Functions for generating a project from a project template."""
-from __future__ import unicode_literals
-
 import fnmatch
-import io
 import json
 import logging
 import os
 import shutil
+import warnings
 from collections import OrderedDict
-
+from pathlib import Path
 from binaryornot.check import is_binary
-from jinja2 import FileSystemLoader
+from jinja2 import FileSystemLoader, Environment
 from jinja2.exceptions import TemplateSyntaxError, UndefinedError
 
-from .environment import StrictEnvironment
-from .exceptions import (
-    NonTemplatedInputDirException,
+from cookiecutter.environment import StrictEnvironment
+from cookiecutter.exceptions import (
     ContextDecodingException,
     FailedHookException,
+    NonTemplatedInputDirException,
     OutputDirExistsException,
-    UndefinedVariableInTemplate
+    UndefinedVariableInTemplate,
 )
-from .find import find_template
-from .hooks import run_hook
-from .utils import make_sure_path_exists, work_in, rmtree
+from cookiecutter.find import find_template
+from cookiecutter.hooks import run_hook
+from cookiecutter.utils import make_sure_path_exists, rmtree, work_in
 
 from .context import context_is_version_2
 
@@ -69,6 +65,15 @@ def apply_overwrites_to_context(context, overwrite_context):
                 # see ``cookiecutter.prompt.prompt_choice_for_config``
                 context_value.remove(overwrite)
                 context_value.insert(0, overwrite)
+            else:
+                raise ValueError(
+                    f"{overwrite} provided for choice variable {variable}, "
+                    f"but the choices are {context_value}."
+                )
+        elif isinstance(context_value, dict) and isinstance(overwrite, dict):
+            # Partially overwrite some keys in original dict
+            apply_overwrites_to_context(context_value, overwrite)
+            context[variable] = context_value
         else:
             # Simply overwrite the value for this variable
             context[variable] = overwrite
@@ -271,8 +276,9 @@ def apply_overwrites_to_context_v2(context, extra_context):
         raise ValueError(msg)
 
 
-def generate_context(context_file='cookiecutter.json', default_context=None,
-                     extra_context=None):
+def generate_context(
+    context_file='cookiecutter.json', default_context=None, extra_context=None
+):
     """Generate the context for a Cookiecutter project template.
 
     Loads the JSON file as a Python object, with key being the JSON filename.
@@ -282,10 +288,10 @@ def generate_context(context_file='cookiecutter.json', default_context=None,
     :param default_context: Dictionary containing config to take into account.
     :param extra_context: Dictionary containing configuration overrides
     """
-    context = {}
+    context = OrderedDict([])
 
     try:
-        with open(context_file) as file_handle:
+        with open(context_file, encoding='utf-8') as file_handle:
             obj = json.load(file_handle, object_pairs_hook=OrderedDict)
     except ValueError as e:
         # JSON decoding error.  Let's throw a new exception that is more
@@ -293,9 +299,10 @@ def generate_context(context_file='cookiecutter.json', default_context=None,
         full_fpath = os.path.abspath(context_file)
         json_exc_message = str(e)
         our_exc_message = (
-            'JSON decoding error while loading "{0}".  Decoding'
-            ' error details: "{1}"'.format(full_fpath, json_exc_message))
-        raise ContextDecodingException(our_exc_message)
+            f"JSON decoding error while loading '{full_fpath}'. "
+            f"Decoding error details: '{json_exc_message}'"
+        )
+        raise ContextDecodingException(our_exc_message) from e
 
     # Add the Python object to the context dictionary
     file_name = os.path.split(context_file)[1]
@@ -308,22 +315,28 @@ def generate_context(context_file='cookiecutter.json', default_context=None,
         logger.debug("Context is version 2")
 
         if default_context:
-            apply_overwrites_to_context_v2(obj, default_context)
+            try:
+                apply_overwrites_to_context_v2(obj, default_context)
+            except ValueError as error:
+                warnings.warn(f"Invalid default received: {error}")
         if extra_context:
             apply_overwrites_to_context_v2(obj, extra_context)
     else:
         logger.debug("Context is version 1")
 
         if default_context:
-            apply_overwrites_to_context(obj, default_context)
+            try:
+                apply_overwrites_to_context(obj, default_context)
+            except ValueError as error:
+                warnings.warn(f"Invalid default received: {error}")
         if extra_context:
             apply_overwrites_to_context(obj, extra_context)
 
-    logger.debug('Context generated is {}'.format(context))
+    logger.debug('Context generated is %s', context)
     return context
 
 
-def generate_file(project_dir, infile, context, env):
+def generate_file(project_dir, infile, context, env, skip_if_file_exists=False):
     """Render filename of infile as name of outfile, handle infile correctly.
 
     Dealing with infile appropriately:
@@ -344,7 +357,7 @@ def generate_file(project_dir, infile, context, env):
     :param context: Dict for populating the cookiecutter's variables.
     :param env: Jinja2 template execution environment.
     """
-    logger.debug('Processing file {}'.format(infile))
+    logger.debug('Processing file %s', infile)
 
     # Render the path to the output file (not including the root project dir)
     outfile_tmpl = env.from_string(infile)
@@ -352,18 +365,19 @@ def generate_file(project_dir, infile, context, env):
     outfile = os.path.join(project_dir, outfile_tmpl.render(**context))
     file_name_is_empty = os.path.isdir(outfile)
     if file_name_is_empty:
-        logger.debug('The resulting file name is empty: {0}'.format(outfile))
+        logger.debug('The resulting file name is empty: %s', outfile)
         return
 
-    logger.debug('Created file at {0}'.format(outfile))
+    if skip_if_file_exists and os.path.exists(outfile):
+        logger.debug('The resulting file already exists: %s', outfile)
+        return
+
+    logger.debug('Created file at %s', outfile)
 
     # Just copy over binary files. Don't render.
-    logger.debug("Check {} to see if it's a binary".format(infile))
+    logger.debug("Check %s to see if it's a binary", infile)
     if is_binary(infile):
-        logger.debug(
-            'Copying binary {} to {} without rendering'
-            ''.format(infile, outfile)
-        )
+        logger.debug('Copying binary %s to %s without rendering', infile, outfile)
         shutil.copyfile(infile, outfile)
     else:
         # Force fwd slashes on Windows for get_template
@@ -380,40 +394,52 @@ def generate_file(project_dir, infile, context, env):
             raise
         rendered_file = tmpl.render(**context)
 
-        logger.debug('Writing contents to file {}'.format(outfile))
+        # Detect original file newline to output the rendered file
+        # note: newline='' ensures newlines are not converted
+        with open(infile, encoding='utf-8', newline='') as rd:
+            rd.readline()  # Read the first line to load 'newlines' value
 
-        with io.open(outfile, 'w', encoding='utf-8') as fh:
+            # Use `_new_lines` overwrite from context, if configured.
+            newline = rd.newlines
+            if context['cookiecutter'].get('_new_lines', False):
+                newline = context['cookiecutter']['_new_lines']
+                logger.debug('Overwriting end line character with %s', newline)
+
+        logger.debug('Writing contents to file %s', outfile)
+
+        with open(outfile, 'w', encoding='utf-8', newline=newline) as fh:
             fh.write(rendered_file)
 
     # Apply file permissions to output file
     shutil.copymode(infile, outfile)
 
 
-def render_and_create_dir(dirname, context, output_dir, environment,
-                          overwrite_if_exists=False):
+def render_and_create_dir(
+    dirname: str,
+    context: dict,
+    output_dir: "os.PathLike[str]",
+    environment: Environment,
+    overwrite_if_exists: bool = False,
+):
     """Render name of a directory, create the directory, return its path."""
     name_tmpl = environment.from_string(dirname)
     rendered_dirname = name_tmpl.render(**context)
 
-    dir_to_create = os.path.normpath(
-        os.path.join(output_dir, rendered_dirname)
+    dir_to_create = Path(output_dir, rendered_dirname)
+
+    logger.debug(
+        'Rendered dir %s must exist in output_dir %s', dir_to_create, output_dir
     )
 
-    logger.debug('Rendered dir {} must exist in output_dir {}'.format(
-        dir_to_create,
-        output_dir
-    ))
-
-    output_dir_exists = os.path.exists(dir_to_create)
+    output_dir_exists = dir_to_create.exists()
 
     if output_dir_exists:
         if overwrite_if_exists:
             logger.debug(
-                'Output directory {} already exists,'
-                'overwriting it'.format(dir_to_create)
+                'Output directory %s already exists, overwriting it', dir_to_create
             )
         else:
-            msg = 'Error: "{}" directory already exists'.format(dir_to_create)
+            msg = f'Error: "{dir_to_create}" directory already exists'
             raise OutputDirExistsException(msg)
     else:
         make_sure_path_exists(dir_to_create)
@@ -429,8 +455,9 @@ def ensure_dir_is_templated(dirname):
         raise NonTemplatedInputDirException
 
 
-def _run_hook_from_repo_dir(repo_dir, hook_name, project_dir, context,
-                            delete_project_on_failure):
+def _run_hook_from_repo_dir(
+    repo_dir, hook_name, project_dir, context, delete_project_on_failure
+):
     """Run hook from repo directory, clean project directory if hook fails.
 
     :param repo_dir: Project template input directory.
@@ -447,14 +474,22 @@ def _run_hook_from_repo_dir(repo_dir, hook_name, project_dir, context,
             if delete_project_on_failure:
                 rmtree(project_dir)
             logger.error(
-                "Stopping generation because {} hook "
-                "script didn't exit successfully".format(hook_name)
+                "Stopping generation because %s hook "
+                "script didn't exit successfully",
+                hook_name,
             )
             raise
 
 
-def generate_files(repo_dir, context=None, output_dir='.',
-                   overwrite_if_exists=False):
+def generate_files(
+    repo_dir,
+    context=None,
+    output_dir='.',
+    overwrite_if_exists=False,
+    skip_if_file_exists=False,
+    accept_hooks=True,
+    keep_project_on_failure=False,
+):
     """Render the templates and saves them to files.
 
     :param repo_dir: Project template input directory.
@@ -462,28 +497,28 @@ def generate_files(repo_dir, context=None, output_dir='.',
     :param output_dir: Where to output the generated project dir into.
     :param overwrite_if_exists: Overwrite the contents of the output directory
         if it exists.
+    :param skip_if_file_exists: Skip the files in the corresponding directories
+        if they already exist
+    :param accept_hooks: Accept pre and post hooks if set to `True`.
+    :param keep_project_on_failure: If `True` keep generated project directory even when
+        generation fails
     """
     template_dir = find_template(repo_dir)
-    logger.debug('Generating project from {}...'.format(template_dir))
-    context = context or {}
+    logger.debug('Generating project from %s...', template_dir)
+    context = context or OrderedDict([])
+
+    envvars = context.get('cookiecutter', {}).get('_jinja2_env_vars', {})
 
     unrendered_dir = os.path.split(template_dir)[1]
     ensure_dir_is_templated(unrendered_dir)
-    env = StrictEnvironment(
-        context=context,
-        keep_trailing_newline=True,
-    )
+    env = StrictEnvironment(context=context, keep_trailing_newline=True, **envvars)
     try:
         project_dir, output_directory_created = render_and_create_dir(
-            unrendered_dir,
-            context,
-            output_dir,
-            env,
-            overwrite_if_exists
+            unrendered_dir, context, output_dir, env, overwrite_if_exists
         )
     except UndefinedError as err:
-        msg = "Unable to create project directory '{}'".format(unrendered_dir)
-        raise UndefinedVariableInTemplate(msg, err, context)
+        msg = f"Unable to create project directory '{unrendered_dir}'"
+        raise UndefinedVariableInTemplate(msg, err, context) from err
 
     # We want the Jinja path and the OS paths to match. Consequently, we'll:
     #   + CD to the template folder
@@ -493,22 +528,19 @@ def generate_files(repo_dir, context=None, output_dir='.',
     # absolute path for the target folder (project_dir)
 
     project_dir = os.path.abspath(project_dir)
-    logger.debug('Project directory is {}'.format(project_dir))
+    logger.debug('Project directory is %s', project_dir)
 
     # if we created the output directory, then it's ok to remove it
     # if rendering fails
-    delete_project_on_failure = output_directory_created
+    delete_project_on_failure = output_directory_created and not keep_project_on_failure
 
-    _run_hook_from_repo_dir(
-        repo_dir,
-        'pre_gen_project',
-        project_dir,
-        context,
-        delete_project_on_failure
-    )
+    if accept_hooks:
+        _run_hook_from_repo_dir(
+            repo_dir, 'pre_gen_project', project_dir, context, delete_project_on_failure
+        )
 
     with work_in(template_dir):
-        env.loader = FileSystemLoader('.')
+        env.loader = FileSystemLoader(['.', '../templates'])
 
         for root, dirs, files in os.walk('.'):
             # We must separate the two types of dirs into different lists.
@@ -523,6 +555,7 @@ def generate_files(repo_dir, context=None, output_dir='.',
                 # specified in the ``_copy_without_render`` setting, but
                 # we store just the dir name
                 if is_copy_only_path(d_, context):
+                    logger.debug('Found copy only path %s', d)
                     copy_dirs.append(d)
                 else:
                     render_dirs.append(d)
@@ -530,10 +563,14 @@ def generate_files(repo_dir, context=None, output_dir='.',
             for copy_dir in copy_dirs:
                 indir = os.path.normpath(os.path.join(root, copy_dir))
                 outdir = os.path.normpath(os.path.join(project_dir, indir))
-                logger.debug(
-                    'Copying dir {} to {} without rendering'
-                    ''.format(indir, outdir)
-                )
+                outdir = env.from_string(outdir).render(**context)
+                logger.debug('Copying dir %s to %s without rendering', indir, outdir)
+
+                # The outdir is not the root dir, it is the dir which marked as copy
+                # only in the config file. If the program hits this line, which means
+                # the overwrite_if_exists = True, and root dir exists
+                if os.path.isdir(outdir):
+                    shutil.rmtree(outdir)
                 shutil.copytree(indir, outdir)
 
             # We mutate ``dirs``, because we only want to go through these
@@ -543,18 +580,14 @@ def generate_files(repo_dir, context=None, output_dir='.',
                 unrendered_dir = os.path.join(project_dir, root, d)
                 try:
                     render_and_create_dir(
-                        unrendered_dir,
-                        context,
-                        output_dir,
-                        env,
-                        overwrite_if_exists
+                        unrendered_dir, context, output_dir, env, overwrite_if_exists
                     )
                 except UndefinedError as err:
                     if delete_project_on_failure:
                         rmtree(project_dir)
                     _dir = os.path.relpath(unrendered_dir, output_dir)
-                    msg = "Unable to create directory '{}'".format(_dir)
-                    raise UndefinedVariableInTemplate(msg, err, context)
+                    msg = f"Unable to create directory '{_dir}'"
+                    raise UndefinedVariableInTemplate(msg, err, context) from err
 
             for f in files:
                 infile = os.path.normpath(os.path.join(root, f))
@@ -563,26 +596,28 @@ def generate_files(repo_dir, context=None, output_dir='.',
                     outfile_rendered = outfile_tmpl.render(**context)
                     outfile = os.path.join(project_dir, outfile_rendered)
                     logger.debug(
-                        'Copying file {} to {} without rendering'
-                        ''.format(infile, outfile)
+                        'Copying file %s to %s without rendering', infile, outfile
                     )
                     shutil.copyfile(infile, outfile)
                     shutil.copymode(infile, outfile)
                     continue
                 try:
-                    generate_file(project_dir, infile, context, env)
+                    generate_file(
+                        project_dir, infile, context, env, skip_if_file_exists
+                    )
                 except UndefinedError as err:
                     if delete_project_on_failure:
                         rmtree(project_dir)
-                    msg = "Unable to create file '{}'".format(infile)
-                    raise UndefinedVariableInTemplate(msg, err, context)
+                    msg = f"Unable to create file '{infile}'"
+                    raise UndefinedVariableInTemplate(msg, err, context) from err
 
-    _run_hook_from_repo_dir(
-        repo_dir,
-        'post_gen_project',
-        project_dir,
-        context,
-        delete_project_on_failure
-    )
+    if accept_hooks:
+        _run_hook_from_repo_dir(
+            repo_dir,
+            'post_gen_project',
+            project_dir,
+            context,
+            delete_project_on_failure,
+        )
 
     return project_dir
